Migrate to typescript
