+++ .prettierrc
@@ -1,0 +1,1 @@
+"@vtex/prettier-config"
