+++ Dockerfile
@@ -1,6 +1,5 @@
 FROM sonarsource/sonar-scanner-cli
 
-COPY . /
-RUN npm install --global yarn
-RUN yarn
-ENTRYPOINT ["node", "/lib/index.js"]
+COPY ./dist /dist
+
+ENTRYPOINT ["node", "/dist/index.js"]
