+++ package.json
@@ -5,7 +5,21 @@
   "repository": "v:action-app-store-review",
   "author": "Thayanne Sousa <thayannevls@gmail.com>",
   "license": "MIT",
-  "devDependencies": {},
+  "scripts": {
+    "build": "ncc build src/index.ts --source-map --license licenses.txt --out dist",
+    "all": "yarn build && yarn package",
+    "lint": "eslint src/",
+    "format": "prettier --write \"src/**/*.{ts,js,json}\""
+  },
+  "devDependencies": {
+    "@vercel/ncc": "^0.27.0",
+    "@vtex/prettier-config": "^0.3.6",
+    "@vtex/tsconfig": "^0.5.6",
+    "eslint": "^7.18.0",
+    "eslint-config-vtex": "^12.9.3",
+    "prettier": "^2.2.1",
+    "typescript": "^4.1.3"
+  },
   "dependencies": {
     "@actions/core": "^1.2.6",
     "@actions/exec": "^1.0.4",
