+++ tsconfig.json
@@ -1,0 +1,19 @@
+{
+	"extends": "@vtex/tsconfig",
+	"compilerOptions": {
+		"types": [
+			"node"
+		],
+		"target": "es2017",
+		"module": "commonjs",
+		"moduleResolution": "node",
+		"esModuleInterop": true,
+		"outDir": "dist",
+	},
+	"include": [
+		"src/**/*.ts"
+	],
+	"exclude": [
+		"node_modules"
+	]
+}
